+++ include/protocol_pro_2.hpp
@@ -65,7 +65,7 @@ class RoverRobotics::Pro2ProtocolObject
    * @param std::vector<uin32_t> Bytes stream from the robot
    * @return structure of statusData
    */
-  void unpack_comm_response(std::vector<uint32_t>) override;
+  void unpack_comm_response(std::vector<uint8_t>) override;
   /*
    * @brief Check if Communication still exist
    * @return bool
